Move Maps Database to Partial, Add 3 Map Setup Types (path, provider, index). Add Thumbnail Component
