+++ Runtime/Resources/Map/Map.Database.cs
@@ -1,0 +1,47 @@
+﻿using System.Collections.Generic;
+
+namespace Espionage.Engine.Resources
+{
+	public sealed partial class Map
+	{
+		private static IDatabase<Map, string> Database { get; } = new InternalDatabase();
+
+		private class InternalDatabase : IDatabase<Map, string>
+		{
+			public IEnumerable<Map> All => _records.Values;
+			public int Count => _records.Count;
+
+			public Map this[ string key ] => _records.ContainsKey( key ) ? _records[key] : null;
+
+			private readonly Dictionary<string, Map> _records = new();
+
+			public void Add( Map item )
+			{
+				// Store it in Database
+				if ( _records.ContainsKey( item.Identifier! ) )
+				{
+					_records[item.Identifier] = item;
+				}
+				else
+				{
+					_records.Add( item.Identifier!, item );
+				}
+			}
+
+			public void Clear()
+			{
+				_records.Clear();
+			}
+
+			public bool Contains( Map item )
+			{
+				return _records.ContainsKey( item.Identifier );
+			}
+
+			public void Remove( Map item )
+			{
+				_records.Remove( item.Identifier );
+			}
+		}
+	}
+}
